+++ package.json
@@ -28,8 +28,9 @@
     "dotenv": "^16.0.3",
     "express": "^4.18.2",
     "http": "^0.0.1-security",
+    "mina-arena-contracts": "^0.1.0",
     "pg": "^8.9.0",
-    "sequelize": "^6.28.0"
+    "sequelize": "^6.31.0"
   },
   "type": "module",
   "devDependencies": {
