+++ src/graphql/mutations/create_game_piece_actions.ts
@@ -2,8 +2,10 @@ import * as Types from '../__generated__/resolvers-types';
 import * as Models from '../../models/index.js';
 import sequelizeConnection from '../../db/config.js';
 import { unique, snakeToCamel, enforceOneOf } from '../helpers.js';
-import { MELEE_ATTACK_RANGE, RANGED_ATTACK_RANGE } from '../../models/unit.js';
-import { resolveMoveAction, resolveRangedAttackAction, resolveMeleeAttackAction } from '../../service_objects/game_piece_action_resolver.js';
+
+import resolveMoveAction from '../../service_objects/game_piece_action_resolvers/move_resolver';
+import resolveRangedAttackAction from '../../service_objects/game_piece_action_resolvers/ranged_attack_resolver';
+import resolveMeleeAttackAction from '../../service_objects/game_piece_action_resolvers/melee_attack_resolver';
 
 export default async (
   parent,
