+++ src/service_objects/game_piece_action_resolvers/move_resolver.ts
@@ -1,0 +1,43 @@
+import * as Models from '../../models/index.js';
+import { GamePieceCoordinates } from '../../graphql/__generated__/resolvers-types.js';
+
+export default async function resolveMoveAction(
+  gamePiece: Models.GamePiece,
+  moveFrom: GamePieceCoordinates,
+  moveTo: GamePieceCoordinates,
+  commitChanges: boolean = false,
+  transaction
+): Promise<Models.GamePiece> {
+  const currentPos = gamePiece.coordinates();
+
+  if (moveFrom.x != currentPos.x || moveFrom.y != currentPos.y) {
+    throw new Error(
+      `GamePiece ${gamePiece.id} is at ${JSON.stringify(currentPos)} ` +
+      `but you are attempting to move it from ${JSON.stringify(moveFrom)}`
+    );
+  }
+  
+  const moveValidityResult = await gamePiece.checkMoveValidity(moveTo);
+  if (!moveValidityResult.valid) {
+    if (moveValidityResult.invalidReason == 'beyondMaxRange') {
+      throw new Error(
+        `GamePiece ${gamePiece.id} cannot be moved from ${JSON.stringify(currentPos)} to ${JSON.stringify(moveTo)} ` +
+        `because the distance is ${moveValidityResult.distance} its movement speed is ${moveValidityResult.movementSpeed}`
+      );
+    }
+    if (moveValidityResult.invalidReason == 'collidesWithOtherPiece') {
+      throw new Error(
+        `GamePiece ${gamePiece.id} cannot be moved from ${JSON.stringify(currentPos)} ` +
+        `to ${JSON.stringify(moveTo)} because this collides with another GamePiece`
+      );
+    }
+  }
+
+  if (commitChanges) {
+    // Validations done, modify state
+    gamePiece.positionX = moveTo.x;
+    gamePiece.positionY = moveTo.y;
+    await gamePiece.save({ transaction: transaction });
+  }
+  return gamePiece;
+}
